cleanup BaseVideoPreviewView
