+++ Wire-iOS/Sources/UserInterface/Calling/VideoGridView/Stream/Stream.swift
