+++ Wire-iOS/Sources/UserInterface/Calling/VideoGridView/Stream/VideoStream.swift
